requirements.txt
Flask==3.0.3
requests==2.26.0
python-dotenv
flask-cors